feat: migrate frontend components and pages to TypeScript
- Converted multiple components and pages from JavaScript to TypeScript, enhancing type safety and maintainability.
- Introduced new TypeScript interfaces for user, event, and report structures to standardize data handling.
- Updated Jest configuration and test files to support TypeScript, ensuring all tests are compatible with the new structure.
- Refactored existing components to utilize the new type definitions and improve overall code quality.
- Removed outdated JavaScript files and streamlined imports across the application for consistency.
- Enhanced the TypeScript migration checklist to reflect the current progress and future tasks.
