--- frontend/context/ModalContext.js
@@ -1,2 +1,0 @@
-import { createContext } from 'react';
-export const ModalContext = createContext({ openModal: () => {} });
