+++ frontend/context/ModalContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export interface ModalContextType {
+  openModal: (...args: any[]) => void;
+}
+
+export const ModalContext = createContext<ModalContextType>({ openModal: () => {} });
