--- frontend/pages/event/[event-slug]/admin/reports/user/[user-id].js
@@ -1,3 +1,0 @@
-import React from "react";
-
-
