--- frontend/pages/event/[event-slug]/user/[user-id]/index.js
@@ -1,3 +1,0 @@
-import React from "react";
-
-
