--- frontend/pages/event/[event-slug]/user/index.js
@@ -1,3 +1,0 @@
-import React from "react";
-
-
