+++ frontend/pages/event/[event-slug]/user/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+// This file is intentionally left minimal, as it was in the JavaScript version
+// Will be implemented in the future
+
+export default function EventUserIndexPage() {
+  return (
+    <div className="p-6">
+      <h1 className="text-xl font-bold">User Events Page</h1>
+      <p className="text-gray-600">This page is under development</p>
+    </div>
+  );
+}
