+++ frontend/pages/login.test.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import LoginPage from './login';
+import { UserContext } from './_app';
+import '@testing-library/jest-dom';
 
+// Mock the Next.js router
 jest.mock('next/router', () => ({
   useRouter: () => ({
     push: jest.fn(),
@@ -13,7 +16,11 @@ jest.mock('next/router', () => ({
 
 describe('Login page', () => {
   it('renders email and password fields and login button', () => {
-    render(<LoginPage />);
+    render(
+      <UserContext.Provider value={{ user: null, setUser: jest.fn() }}>
+        <LoginPage />
+      </UserContext.Provider>
+    );
     expect(screen.getByLabelText(/email/i)).toBeInTheDocument();
     expect(screen.getByLabelText(/password/i)).toBeInTheDocument();
     expect(screen.getByRole('button', { name: /login/i })).toBeInTheDocument();
